+++ R/mod_common_forms.R
@@ -11,9 +11,9 @@ mod_common_forms_ui <- function(id, form){
     bslib::layout_sidebar(
       fillable = FALSE,
       if(form == "Adverse events"){
-        DT::dataTableOutput(ns("SAE_table"))
+        mod_review_form_tbl_ui(ns("review_form_SAE_tbl"))
       },
-      DT::dataTableOutput(ns("common_form_table")), 
+      mod_review_form_tbl_ui(ns("review_form_tbl")),
       sidebar = bslib::sidebar(
         bg = "white", 
         position = "right",
@@ -73,7 +73,8 @@ mod_common_forms_ui <- function(id, form){
 #'   interactive tables.
 #'
 #' @seealso [mod_common_forms_ui()], [mod_timeline_ui()],
-#'   [mod_timeline_server()]
+#'   [mod_timeline_server()], [mod_review_form_tbl_ui()],
+#'   [mod_review_form_tbl_server()]
 #' 
 mod_common_forms_server <- function(
     id, 
@@ -93,159 +94,62 @@ mod_common_forms_server <- function(
   moduleServer( id, function(input, output, session){
     ns <- session$ns
 
-    data_active <- reactive({
-      req(!is.null(input$show_all_data))
-      shiny::validate(need(
-        !is.null(r$filtered_data[[form]]),
-        paste0("Warning: no data found in the database for the form '", form, "'.")
-      ))
-      df <- dplyr::left_join(
-        r$filtered_data[[form]],
-        with(r$review_data, r$review_data[item_group == form, ]) |> 
-          dplyr::select(-dplyr::all_of(c("edit_date_time", "event_date"))), 
-        by = id_item
-      ) |> 
-        dplyr::mutate(
-          item_value = ifelse(
-            reviewed == "No", 
-            paste0("<b>", htmltools::htmlEscape(item_value), "*</b>"), 
-            htmltools::htmlEscape(item_value)
-          )
+    common_form_data <- reactiveVal()
+    SAE_data <- reactiveVal()
+    observe({
+      df <- {
+        shiny::validate(need(
+          !is.null(r$filtered_data[[form]]),
+          paste0("Warning: no data found in the database for the form '", form, "'.")
+        ))
+        dplyr::left_join(
+          r$filtered_data[[form]],
+          with(r$review_data, r$review_data[item_group == form, ]) |> 
+            dplyr::select(-dplyr::all_of(c("edit_date_time", "event_date"))), 
+          by = id_item
         ) |> 
-        create_table(expected_columns = names(form_items))
-      if(!input$show_all_data){ 
-        df <-  with(df, df[subject_id == r$subject_id, ]) 
+          dplyr::mutate(
+            item_value = ifelse(
+              reviewed == "No", 
+              paste0("<b>", htmltools::htmlEscape(item_value), "*</b>"), 
+              htmltools::htmlEscape(item_value)
+            )
+          ) |> 
+          create_table(expected_columns = names(form_items)) |> 
+          dplyr::mutate(o_reviewed = Map(\(x, y, z) append(x, list(row_id = y, disabled = z)), 
+                                         o_reviewed, 
+                                         dplyr::row_number(),
+                                         subject_id != r$subject_id))
       }
-      df
+      common_form_data({
+        if(form == "Adverse events") {
+          df |> 
+            dplyr::filter(!grepl("Yes", `Serious Adverse Event`)
+            ) |> 
+            dplyr::select(-dplyr::starts_with("SAE"))
+        } else {
+          df
+        }
+      })
+      if (form == "Adverse events")
+        SAE_data({
+          df |>
+            dplyr::filter(grepl("Yes", `Serious Adverse Event`)) |>
+            dplyr::select(dplyr::any_of(
+              c("o_reviewed", "subject_id","form_repeat", "Name", "AESI",  "SAE Start date",
+                "SAE End date", "CTCAE severity", "Treatment related",
+                "Treatment action", "Other action", "SAE Category",
+                "SAE Awareness date", "SAE Date of death", "SAE Death reason")
+            )) |>
+            adjust_colnames("^SAE ")
+        })
     })
+    mod_review_form_tbl_server("review_form_tbl", r, common_form_data, form, reactive(input$show_all_data), table_names, form)
+    if (form == "Adverse events")
+      mod_review_form_tbl_server("review_form_SAE_tbl", r, SAE_data, form, reactive(input$show_all_data), table_names, "Serious Adverse Events")
     
     mod_timeline_server("timeline_fig", r = r, form = form)
     
-    output[["SAE_table"]] <- DT::renderDT({
-      req(form == "Adverse events")
-      SAE_data <- data_active() |> 
-        dplyr::filter(grepl("Yes", `Serious Adverse Event`)) |> 
-        dplyr::select(dplyr::any_of(
-          c("o_reviewed", "subject_id","form_repeat", "Name", "AESI",  "SAE Start date", 
-            "SAE End date", "CTCAE severity", "Treatment related", 
-            "Treatment action", "Other action", "SAE Category", 
-            "SAE Awareness date", "SAE Date of death", "SAE Death reason")
-        )) |> 
-        adjust_colnames("^SAE ")
-      if(!input$show_all_data) SAE_data$subject_id <- NULL
-
-      # determine DT dom / exts / opts
-      DT <- dt_config(SAE_data,
-        table_name = paste("SAE", ifelse(input$show_all_data,
-        "all_patients", r$subject_id), sep = ".")) 
-      datatable_custom(
-        SAE_data, 
-        rename_vars = c("Review Status" = "o_reviewed", table_names), 
-        rownames = FALSE,
-        title = "Serious Adverse Events", 
-        escape = FALSE,
-        dom = DT$dom,
-        extensions = DT$exts,
-        selection = "none",
-        callback = checkbox_callback,
-        options = append(list(
-          columnDefs = list(list(
-            targets = 0,
-            render = checkbox_render
-          )),
-          createdRow = checkbox_create_callback
-        ),
-        DT$opts))
-    }, server = FALSE)
-
-    observeEvent(data_active(), {
-      session$userData$update_checkboxes[[form]] <- NULL
-      session$userData$review_records[[form]] <- data.frame(id = integer(), reviewed = character())
-    })
-    
-    observeEvent(input$common_form_table_review_selection, {
-      session$userData$update_checkboxes[[form]] <- NULL
-      
-      session$userData$review_records[[form]] <-
-        dplyr::rows_upsert(
-          session$userData$review_records[[form]],
-          input$common_form_table_review_selection,
-          by = "id"
-        ) |> 
-        dplyr::filter(!is.na(reviewed)) |> 
-        dplyr::semi_join(
-          subset(r$review_data, subject_id == r$subject_id & item_group == form),
-          by = "id"
-        ) |> 
-        dplyr::anti_join(
-          subset(r$review_data, subject_id == r$subject_id & item_group == form),
-          by = c("id", "reviewed")
-        ) |> 
-        dplyr::arrange(id)
-    })
-    
-    observeEvent(session$userData$update_checkboxes[[form]], {
-      checked <- session$userData$update_checkboxes[[form]]
-      
-      update_cbs(ns("common_form_table"), checked)
-      update_cbs(ns("SAE_table"), checked)
-    })
-    
-    observeEvent(input$SAE_table_review_selection, {
-      session$userData$update_checkboxes[[form]] <- NULL
-      
-      session$userData$review_records[[form]] <-
-        dplyr::rows_upsert(
-          session$userData$review_records[[form]],
-          input$SAE_table_review_selection,
-          by = "id"
-        ) |> 
-        dplyr::filter(!is.na(reviewed)) |> 
-        dplyr::semi_join(
-          subset(r$review_data, subject_id == r$subject_id & item_group == form),
-          by = "id"
-        ) |> 
-        dplyr::anti_join(
-          subset(r$review_data, subject_id == r$subject_id & item_group == form),
-          by = c("id", "reviewed")
-        ) |> 
-        dplyr::arrange(id)
-    })
-    
-    output[["common_form_table"]] <- DT::renderDT({
-      df <- data_active()
-      if(form == "Adverse events") {
-        df <- df |> 
-          dplyr::filter(!grepl("Yes", `Serious Adverse Event`)
-          ) |> 
-          dplyr::select(-dplyr::starts_with("SAE"))
-      }
-      if(!input$show_all_data) df$subject_id <- NULL
-
-      # determine DT dom / exts / opts
-      DT <- dt_config(df,
-         table_name = paste(form, ifelse(input$show_all_data,
-         "all_patients", r$subject_id), sep = ".")) 
-      datatable_custom(
-        df, 
-        rename_vars = c("Review Status" = "o_reviewed", table_names), 
-        rownames= FALSE,
-        title = form, 
-        escape = FALSE,
-        dom = DT$dom,
-        extensions = DT$exts,
-        selection = "none",
-        callback = checkbox_callback,
-        options = append(list(
-          columnDefs = list(list(
-            targets = 0,
-            render = checkbox_render
-          )),
-          createdRow = checkbox_create_callback
-        ),
-        DT$opts))
-    }, server = FALSE)
-
   })
 }