+++ NEWS.md
@@ -9,10 +9,15 @@
 - Added form type as a class to be used in `create_table()` to display tables.
 - Add a logging table to the DB for reviews.
 - Simplify pulling data from DB for reviews.
+- Review data by records IDs instead of subject & form
+- Make query handling a configurable option
+- Changed the legend to display 'significance pending' instead of 'significance unknown'.
+- Added `Excel` download button to Queries table & patient listings that need review.
 
 ## Bug fixes
 
 - The test-coverage GHA workflow is updated so that codecov uploads work again.
+- Display all rows for tables where `Scroller` is disabled.
 
 ## `devex` changes
 - Added `Excel` download button to Queries table & patient listings that need review.